requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
cython==3.0.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
django==5.1.5; python_version >= '3.10'
django-ckeditor-5==0.2.15; python_version >= '3.7'
djangorestframework==3.15.2; python_version >= '3.8'
joblib==1.4.2; python_version >= '3.8'
numpy==2.2.2; python_version >= '3.10'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
patsy==1.0.1; python_version >= '3.6'
pillow==11.1.0; python_version >= '3.9'
pmdarima==2.0.4; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2025.1
scikit-learn==1.6.1; python_version >= '3.9'
scipy==1.15.1; python_version >= '3.10'
setuptools==75.8.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.3; python_version >= '3.8'
statsmodels==0.14.4; python_version >= '3.9'
threadpoolctl==3.5.0; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
urllib3==2.3.0; python_version >= '3.9'
googletrans==4.0.0-rc1

django-redis>=5.0.0
